fix(sgwc): print pool sizes with %llu cast; PRIu64 lacked inttypes.h

+++ src/sgwc/sgwc-workers.c
@@ -361,10 +361,10 @@ int sgwc_workers_start(void)
     sgwc_worker_count = sgwc_worker_configured;
 
     ogs_info("SGWC SMP workers: %d shard(s), shared UE/session pools "
-            "(max.ue=%" PRIu64 " sess=%" PRIu64 ")",
+            "(max.ue=%llu sess=%llu)",
             sgwc_worker_count,
-            (uint64_t)ogs_global_conf()->max.ue,
-            (uint64_t)ogs_app()->pool.sess);
+            (unsigned long long)ogs_global_conf()->max.ue,
+            (unsigned long long)ogs_app()->pool.sess);
 
     return OGS_OK;
 }
